fix(dev): rename Analytics module to Insights to dodge content blockers
Content/tracker blockers block any URL whose path contains "analytics".
The Vite dev server serves each source file at its filesystem path, so the
root layout's static import of $lib/Analytics fetched
/src/lib/Analytics/index.js — which blockers reject, causing SvelteKit's
client start() to throw "Importing a module script failed" and leaving the
app stuck on "Loading…". Safari-only (that's where the blocker lives) and
dev-only (production ships hash-named chunks with no "analytics" in the path).

Rename src/lib/Analytics → src/lib/Insights (and Analytics.js/analyticsUtils.js
→ Insights.js/insightsUtils.js) so dev-server paths no longer match blocklists.
Only file paths changed: the Analytics class name, the `analytics` import
identifiers, and the PUBLIC_ANALYTICS_PROVIDER env var are all unchanged.
