+++ src/lib/Insights/adapters/NoopAdapter.js
