+++ src/lib/Insights/adapters/PlausibleAdapter.js
