+++ src/lib/Insights/adapters/UmamiAdapter.js
