+++ src/lib/Insights/createAdapter.js
