+++ src/lib/Insights/insightsUtils.js
